@@ -117,40 +117,46 @@
     "eslint-plugin-import": "~2.8.0",
     "eslint-plugin-jsx-a11y": "~6.0.3",
     "eslint-plugin-react": "~7.5.1",
-    "extract-text-webpack-plugin": "~3.0.2",
-    "file-loader": "~1.1.6",
+    "file-loader": "^4.2.0",
     "happypack": "^5.0.0",
-    "html-webpack-plugin": "~2.30.1",
+    "html-webpack-plugin": "^3.2.0",
     "http-proxy": "~1.16.2",
     "jest": "^23.6.0",
     "jsdom": "^12.2.0",
     "less": "~2.7.1",
-    "less-loader": "~4.0.5",
+    "less-loader": "^4.0.6",
+    "mini-css-extract-plugin": "^0.8.0",
     "mockjs": "^1.0.1-beta3",
-    "open-browser-webpack-plugin": "~0.0.2",
-    "postcss-loader": "~2.0.10",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "postcss-loader": "^3.0.0",
     "react-addons-test-utils": "~15.6.2",
-    "react-hot-loader": "^3.0.0-beta.7",
     "sinon": "~4.1.6",
     "socket.io": "2.0.4",
-    "style-loader": "~0.19.1",
+    "style-loader": "^0.19.1",
     "supervisor": "^0.12.0",
-    "url-loader": "~0.6.2",
-    "webpack-bundle-analyzer": "^2.9.0",
-    "webpack-dev-server": "~2.11.0",
+    "terser-webpack-plugin": "^2.2.1",
+    "url-loader": "^2.2.0",
+    "webpack": "^4.41.2",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0",
     "webpack-merge": "^4.1.1",
-    "webpack-parallel-uglify-plugin": "^1.1.0"
+    "webpack-parallel-uglify-plugin": "^1.1.2"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.5",
-    "babel-polyfill": "~6.26.0",
+    "@babel/core": "^7.7.2",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/polyfill": "^7.7.0",
+    "@babel/runtime": "^7.7.2",
+    "@babel/runtime-corejs2": "^7.7.2",
+    "antd": "^3.25.2",
+    "axios": "^0.19.0",
     "benz-amr-recorder": "^1.0.9",
     "clipboard": "^2.0.4",
+    "core-js": "^2.6.10",
     "crypto-js": "^3.1.9-1",
     "draft-js": "^0.10.3",
-    "echarts": "^4.0.2",
+    "echarts": "^4.5.0",
     "enzyme-adapter-react-16": "^1.6",
     "esri-loader": "0.3.1",
     "husky": "^1.1.2",
@@ -158,10 +164,10 @@
     "immutable": "~3.8.1",
     "lint-staged": "^7.3.0",
     "md5": "^2.2.1",
-    "prop-types": "^15.6.2",
+    "prop-types": "^15.7.2",
     "rc-queue-anim": "^1.4.1",
     "rc-tween-one": "~1.7.3",
-    "react": "^16.5.2",
+    "react": "^16.12.0",
     "react-copy-to-clipboard": "~5.0.1",
     "react-dom": "^16.5.2",
     "react-redux": "~5.0.6",
