@@ -22,7 +22,7 @@ export default function configure(initialState) {
   const store = createStoreWithMiddleware(rootReducer, initialState)
 
   if (module.hot) {
-    module.hot.accept('../reducers', () => {
+    module.hot.accept('@reducers', () => {
       store.replaceReducer(nextReducer)
     })
   }
