@@ -1,17 +1,17 @@
 const webpack = require('webpack')
 const path = require('path')
-const ParallelUglifyPlugin = require('webpack-parallel-uglify-plugin')
-const CleanWebpackPlugin = require('clean-webpack-plugin')
-const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer')
+const { CleanWebpackPlugin } = require('clean-webpack-plugin')
 
 module.exports = {
+  mode:"production",
   entry: {
     vendor: [
-      'babel-polyfill',
+      // '@babel/polyfill',
       'react', 
       'react-dom', 
       'react-router', 
       'axios',
+      'lodash'
     ],
     redux: [
       'redux',
@@ -20,22 +20,24 @@ module.exports = {
       'react-router-redux',
     ],
   },
+  devtool: 'source-mapcheap-module-eval-souce-map',
   output: {
     filename: '[name].dll.js',
     path: path.join(__dirname, '../app/resource/dll'),
     library: '[name]_[hash]', 
   },
+  performance: false,
   plugins: [
+    new CleanWebpackPlugin({
+      cleanOnceBeforeBuildPatterns:[path.join(__dirname, '../app/resource/dll')],
+      verbose:true,
+    }),
     // 定义环境变量为开发环境
     new webpack.DefinePlugin({
       'process.env.NODE_ENV': JSON.stringify('production'),
       IS_DEVELOPMETN: true,
     }),
-    new CleanWebpackPlugin(['dll'],{
-      root: path.join(__dirname, '../app/resource/dll'),
-      verbose:false,
-      // exclude:['img']//不删除img静态资源
-    }),
+   
     // 使用插件 DllPlugin
     new webpack.DllPlugin({
       path: path.join(__dirname, '../app/resource/dll', '[name].manifest.json'),
@@ -43,20 +45,7 @@ module.exports = {
       name: '[name]_[hash]',
       context: __dirname
     }),
-    /* 多核压缩代码 */
-    new ParallelUglifyPlugin({
-      cacheDir: '.cache/',
-      uglifyJS:{
-        output: {
-          comments: false
-        },
-        compress: {
-          warnings: false
-        }
-      }
-    }),
-		// new BundleAnalyzerPlugin({
-		// 	analyzerMode: 'static'
-		// }),
+    
+
   ]
 };
